ksef2/py/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.2
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
datamodel-code-generator==0.35.0
genson==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflect==7.5.0
isort==6.1.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==10.8.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.0
pydantic-xml==2.17.3
pydantic_core==2.41.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
signxml==4.2.0
sniffio==1.3.1
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xades==1.0.0
xmlsec==1.3.16
xmlsig==1.0.1
